refactor(ramses_tx): extract Engine and ApplicationMessage to enforce SRP
- Split `src/ramses_tx/gateway.py` into `engine.py` and `application_message.py` to adhere to the Single Responsibility Principle.
- Remove the legacy `ramses_tx/gateway.py` module.
- Update module exports in `src/ramses_tx/__init__.py` to expose the new file structures.
- Refactor downstream imports across `ramses_rf` (`gateway.py`, `entity_state.py`, `device/heat.py`).
- Rename `test_gateway_tx.py` to `test_engine.py` (or `test_engine_tx.py`) and update all `unittest.mock.patch` target strings to point to the new module paths.
- Fix hardcoded patch strings in `tests/tests/test_hgi_id.py`.
- Clean up obsolete Mypy `# type: ignore` directives in `ramses_rf/gateway.py` as strict typing is now natively satisfied.

This establishes the clean architectural boundary required to begin Priority 2: Configuration Consolidation.
