+++ src/ramses_tx/__init__.py
@@ -18,6 +18,7 @@ from .address import (
     Address,
     is_valid_dev_id,
 )
+from .application_message import ApplicationMessage
 from .command import CODE_API_MAP, Command
 from .const import (
     DEV_ROLE_MAP,
@@ -42,7 +43,7 @@ from .const import (
     ZoneRole,
 )
 from .discovery import is_hgi80
-from .gateway import Engine
+from .engine import Engine
 from .logger import set_pkt_logging
 from .message import Message
 from .packet import PKT_LOGGER, Packet
@@ -144,6 +145,8 @@ __all__ = [
     #
     "is_valid_dev_id",
     "set_pkt_logging_config",
+    "ApplicationMessage",
+    "Engine",
 ]
